+++ packages/jest-cli/src/__tests__/watch_test_name_pattern_mode.test.js
@@ -13,15 +13,22 @@ import {KEYS} from 'jest-watcher';
 
 const runJestMock = jest.fn();
 
-jest.mock('ansi-escapes', () => ({
-  clearScreen: '[MOCK - clearScreen]',
-  cursorDown: (count = 1) => `[MOCK - cursorDown(${count})]`,
-  cursorHide: '[MOCK - cursorHide]',
-  cursorRestorePosition: '[MOCK - cursorRestorePosition]',
-  cursorSavePosition: '[MOCK - cursorSavePosition]',
-  cursorShow: '[MOCK - cursorShow]',
-  cursorTo: (x, y) => `[MOCK - cursorTo(${x}, ${y})]`,
-}));
+jest
+  .mock('ansi-escapes', () => ({
+    cursorDown: (count = 1) => `[MOCK - cursorDown(${count})]`,
+    cursorHide: '[MOCK - cursorHide]',
+    cursorRestorePosition: '[MOCK - cursorRestorePosition]',
+    cursorSavePosition: '[MOCK - cursorSavePosition]',
+    cursorShow: '[MOCK - cursorShow]',
+    cursorTo: (x, y) => `[MOCK - cursorTo(${x}, ${y})]`,
+  }))
+  .mock('jest-util', () => {
+    const {specialChars, ...util} = jest.requireActual('jest-util');
+    return {
+      ...util,
+      specialChars: {...specialChars, CLEAR: '[MOCK - clear]'},
+    };
+  });
 
 jest.mock(
   '../SearchSource',