+++ packages/jest-watcher/src/PatternPrompt.js
@@ -13,7 +13,10 @@ import type {ScrollOptions} from 'types/Watch';
 
 import chalk from 'chalk';
 import ansiEscapes from 'ansi-escapes';
+import {specialChars} from 'jest-util';
 import Prompt from './lib/Prompt';
+
+const {CLEAR} = specialChars;
 
 const usage = (entity: string) =>
   `\n${chalk.bold('Pattern Mode Usage')}\n` +
@@ -38,7 +41,7 @@ export default class PatternPrompt {
 
   run(onSuccess: Function, onCancel: Function, options?: {header: string}) {
     this._pipe.write(ansiEscapes.cursorHide);
-    this._pipe.write(ansiEscapes.clearScreen);
+    this._pipe.write(CLEAR);
 
     if (options && options.header) {
       this._pipe.write(options.header + '\n');